omg-social-app: some refactor in messagebox and text input message section component
