+++ src/components/message-box/text-input-message-section/TextInputMessageSection.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 
 import { Link } from "react-router-dom";
 import { useSelector } from "react-redux";
@@ -16,7 +16,8 @@ export default function TextInputMessageSection({
   messageText,
   handleTypingMessage,
   loadingState,
-  sendNewMessage,
+    sendNewMessage,
+  setIsTyping,
 }) {
   const currentUserAvatarFromSlice = useSelector(
     (state) => state.user.userAvatarPicture,
@@ -24,6 +25,23 @@ export default function TextInputMessageSection({
   const currentUserIdFromSlice = useSelector((state) => state.user.userId);
 
   const currentUserNameFromSlice = useSelector((state) => state.user.userName);
+
+  useEffect(() => {
+    if (messageText) {
+      setIsTyping(true);
+    } else {
+      setIsTyping(false);
+      return;
+    }
+
+    const timerToStopType = setTimeout(() => {
+      setIsTyping(false);
+    }, 800);
+
+    return () => {
+      clearTimeout(timerToStopType);
+    };
+  }, [messageText, setIsTyping]);
 
   return (
     <div className="send-message-container">
